+++ example/basic.dart
@@ -8,7 +8,7 @@ import 'dart:async';
 
 main() {
   var bar = new ProgressBar(' [:bar] :percent :etas ', total: 10);
-  var timer = new Timer.periodic(new Duration(milliseconds: 500), (Timer timer) {
+  new Timer.periodic(new Duration(milliseconds: 500), (Timer timer) {
     bar.tick();
     if (bar.complete) {
       timer.cancel();