+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## 0.1.0
+## 0.2.0
 
-* Initial version.
+* Fix compile error
 
 ## 0.1.1
 
 * Make options optional.
 * Add the optional callback function, which called when the progress bar completes.
+
+## 0.1.0
+
+* Initial version.